Refactor: Split files by ScreenActivity.
